memblock: Declare scratch memory as CMA
When we finish populating our memory, we don't want to lose the scratch
region as memory we can use for useful data. Do do that, we mark it as
CMA memory. That means that any allocation within it only happens with
movable memory which we can then happily discard for the next kexec.

That way we don't lose the scratch region's memory anymore for
allocations after boot.

Signed-off-by: Alexander Graf <graf@amazon.com>

---

v1 -> v2:

  - test bot warning fix

v2 -> v3:

  - s/reserve_scratch_mem/mark_phys_as_cma/
  - Declare scratch memory as CMA: Remove ifdeffery
  - Declare scratch memory as CMA: Remove superfluous comment