WIP: Expand scratch lifetime
One of the conclusions at Linux Plumbers Conference was to see how
viable it would be to extend the scratch phase from "as short as
possible" to "from early boot until the buddy allocator". By covering
that longer period, we would no longer need to maintain mem maps for
early boot code or update memblocks - which can be slow for large number
of memory allocations.

This patch is incomplete. I started off massaging the code into the
direction above, but did not finish.

On top of the above, we are missing for the next upstream submission:

  - Add KHO support for reserve_mem, it might be an easier initial
    target than ftrace
  - NUMA support
  - Multiple scratch regions
  - Dynamic scratch resizing
  - Try to port an FS to KHO. Hugetlbfs would be a great target IMHO.